7392790435079732049
ManifestID: 7392790435079732049
Date: 12 June 2025 – 16:33:28 UTC
Version: U56-674504

+++ Assembly-CSharp/BaseCrabConfig.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 
 public static class BaseCrabConfig
 {
-	public static GameObject BaseCrab(string id, string name, string desc, string anim_file, string traitId, bool is_baby, string symbolOverridePrefix = null, string onDeathDropID = "CrabShell", int onDeathDropCount = 1)
+	public static GameObject BaseCrab(string id, string name, string desc, string anim_file, string traitId, bool is_baby, string symbolOverridePrefix = null, string onDeathDropID = "CrabShell", float onDeathDropCount = 1f)
 	{
 		float num = 100f;
 		int num2 = (is_baby ? 1 : 2);
